Integrate pure python tensor ops into core
