--- speaktome/domains/pure/__init__.py
@@ -1,1 +1,0 @@
-"""Pure Python domain modules."""
