+++ testing/stub_todo.txt
@@ -1,16 +1,17 @@
 Stub tests remaining:
-- tests/test_all_classes.py::test_class_stub[speaktome.beam_graph_operator-BeamGraphOperator]
-- tests/test_all_classes.py::test_class_stub[speaktome.beam_retirement_manager-BeamRetirementManager]
-- tests/test_all_classes.py::test_class_stub[speaktome.beam_search-BeamSearch]
-- tests/test_all_classes.py::test_class_stub[speaktome.beam_search_instruction-BeamSearchInstruction]
-- tests/test_all_classes.py::test_class_stub[speaktome.beam_tree_node-BeamTreeNode]
-- tests/test_all_classes.py::test_class_stub[speaktome.beam_tree_visualizer-BeamTreeVisualizer]
-- tests/test_all_classes.py::test_class_stub[speaktome.compressed_beam_tree-CompressedBeamTree]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.beam_graph_operator-BeamGraphOperator]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.beam_retirement_manager-BeamRetirementManager]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.beam_search-BeamSearch]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.beam_search_instruction-BeamSearchInstruction]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.beam_tree_node-BeamTreeNode]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.beam_tree_visualizer-BeamTreeVisualizer]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.compressed_beam_tree-CompressedBeamTree]
 - tests/test_all_classes.py::test_class_stub[speaktome.cpu_demo-RandomModel]
-- tests/test_all_classes.py::test_class_stub[speaktome.human_pilot_controller-HumanPilotController]
-- tests/test_all_classes.py::test_class_stub[speaktome.human_scorer_policy_manager-HumanScorerPolicyManager]
-- tests/test_all_classes.py::test_class_stub[speaktome.model_abstraction-PyTorchModelWrapper]
-- tests/test_all_classes.py::test_class_stub[speaktome.pyg_graph_controller-PyGGraphController]
-- tests/test_all_classes.py::test_class_stub[speaktome.pygeo_mind-PyGeoMind]
-- tests/test_all_classes.py::test_class_stub[speaktome.scorer-Scorer]
-- tests/test_all_classes.py::test_class_stub[speaktome.tensor_abstraction-PyTorchTensorOperations]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.human_pilot_controller-HumanPilotController]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.human_scorer_policy_manager-HumanScorerPolicyManager]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.model_abstraction-PyTorchModelWrapper]
+- tests/test_all_classes.py::test_class_stub[speaktome.domains.geo.pyg_graph_controller-PyGGraphController]
+- tests/test_all_classes.py::test_class_stub[speaktome.domains.geo.pygeo_mind-PyGeoMind]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.scorer-Scorer]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.tensor_abstraction-PyTorchTensorOperations]
+- tests/test_all_classes.py::test_class_stub[speaktome.core.tensor_abstraction-PurePythonTensorOperations]
